Move optimization to opt.py.
