+++ theano/sparse/sandbox/sp2.py
@@ -17,19 +17,21 @@ from theano.sparse.basic import (
     Cast, bcast, wcast, icast, lcast, fcast, dcast, ccast, zcast,
     HStack, hstack, VStack, vstack,
     AddSSData, add_s_s_data,
-    MulSDCSC, mul_s_d_csc, MulSDCSR, mul_s_d_csr,
+    MulSV, mul_s_v,
     Multinomial, multinomial, Poisson, poisson,
     Binomial, csr_fbinomial, csc_fbinomial, csr_dbinomial, csc_dbinomial,
     structured_monoid,
     structured_sigmoid, structured_exp, structured_log, structured_pow,
     structured_minimum, structured_maximum, structured_add,
-    MulSV, mul_s_v, MulSVCSR, mul_s_v_csr,
     StructuredAddSV, structured_add_s_v,
-    StructuredAddSVCSR, structured_add_s_v_csr,
-    SamplingDot, sampling_dot, SamplingDotCSR, sampling_dot_csr)
+    SamplingDot, sampling_dot)
 
 # Also for compatibility
 from theano.sparse.opt import (
+    MulSDCSC, mul_s_d_csc, MulSDCSR, mul_s_d_csr,
+    MulSVCSR, mul_s_v_csr,
+    StructuredAddSVCSR, structured_add_s_v_csr,
+    SamplingDotCSR, sampling_dot_csr,
     local_mul_s_d, local_mul_s_v,
     local_structured_add_s_v, local_sampling_dot_csr)
